user-scripts: fix interceptInline calling call()'s return as a function
The first inline-bridge commit wrote `var c = call(); c(HANDLER, src)`,
but `call()` invokes the bridge with its own args and returns the
bridge's result — calling a Promise as a function throws TypeError, so
the very first inline append blew up before the bridge could fire.
Now matches the proven src-intercept pattern: `var r = call(H, s); if
(!r) return null`.

Caught by a new jsdom replication of DarkReader's exact injectProxy
pattern (test/js/user_script_shim.test.js) plus seven sibling
scenarios. Shim template + buildUserScriptShim split into a pure-Dart
file (user_script_shim.dart) so tool/dump_shim_js.dart can reach the
builder under `fvm dart run` without pulling in flutter_inappwebview.
